As I abandoned the attempt at implementing a basic word2vec without tensorflow, I went back to the code of assigment5 from  Vanhouce's DeepLearning minicourse. However I don't know if updates to tensorflow broke that implementation, but I could not resolve a type conflict in an internal Mul operation not explicitly specified in the code. Even though all variables were explicitly initialized with int32 were needed, at some point it tried to multiply with a structure using float 32. The current trump_nlp.ipynb is based on the very similar tutorial from the tensorflow website, https://www.tensorflow.org/tutorials/word2vec, using the vanilla implementation described there; tensorflow/tensorflow/examples/tutorials/word2vec/word2vec_basic.py
Notable changes;
Since the trump text corpus I compiled is orders of magnitude smaller than
the wikipedia dump Vanhouce's class and the tensorflow tutorial use, the
vocabulary size is all words used. Also rare words are not excluded.
The batch size is and embedding I also chose to be smaller only 48. However
the window size is 6 and num_skips is 12. So more context words are being considered
when making the batches. Since the point was to spot some patterns or simply
observe common themes the "validation" set is not random at all, but consists
of the _loaded_ words;

["hillary", "america", "great", "terrorism", "press",  "china", "build", "president"]